Update Go client codegen plugin to follow idiomatic plugin name
See: https://smithy.io/2.0/guides/building-codegen/configuring-the-generator.html#how-to-name-codegen-plugins
