+++ codegen/smithy-go-codegen/src/main/resources/META-INF/services/software.amazon.smithy.build.SmithyBuildPlugin
@@ -1,1 +1,2 @@
+software.amazon.smithy.go.codegen.GoClientCodegenPlugin
 software.amazon.smithy.go.codegen.GoCodegenPlugin
